refactor: #55, #56 import statements in __init__.py, img2thumb.py, and rde2util.py
