+++ src/rdetoolkit/__init__.pyi
@@ -1,6 +1,6 @@
 from .impl import *
 from .interfaces import *
-from . import exceptions as exceptions, img2thumb as img2thumb, invoicefile as invoicefile, modeproc as modeproc, rde2util as rde2util, rdelogger as rdelogger, workflows as workflows
+from . import errors as errors, exceptions as exceptions, img2thumb as img2thumb, invoicefile as invoicefile, modeproc as modeproc, rde2util as rde2util, rdelogger as rdelogger, workflows as workflows
 from .models import config as config, invoice_schema as invoice_schema, metadata as metadata, rde2types as rde2types
 from rdetoolkit.core import resize_image_aspect_ratio as resize_image_aspect_ratio
