+++ src/rdetoolkit/exceptions.pyi
@@ -1,8 +1,5 @@
-import logging
-import traceback
 from _typeshed import Incomplete
-from collections.abc import Generator
-from typing import Any, Callable
+from typing import Any
 
 class StructuredError(Exception):
     emsg: Incomplete
@@ -39,8 +36,6 @@ class RdeFormatModeError(Exception):
     traceback_info: Incomplete
     def __init__(self, emsg: str = '', ecode: int = 103, eobj: Any | None = None, traceback_info: str | None = None) -> None: ...
 
-def skip_exception_context(exception_type: type[Exception], logger: logging.Logger | None = None, enabled: bool = False) -> Generator[dict[str, str | None], None, None]: ...
-
 class InvoiceSchemaValidationError(Exception):
     message: Incomplete
     def __init__(self, message: str = 'Validation error') -> None: ...
@@ -48,7 +43,3 @@ class InvoiceSchemaValidationError(Exception):
 class MetadataValidationError(Exception):
     message: Incomplete
     def __init__(self, message: str = 'Validation error') -> None: ...
-
-def format_simplified_traceback(tb_list: list[traceback.FrameSummary]) -> str: ...
-def handle_exception(e: Exception, error_message: str | None = None, error_code: int | None = None, eobj: Any | None = None, verbose: bool = False) -> StructuredError: ...
-def catch_exception_with_message(*, error_message: str | None = None, error_code: int | None = None, eobj: Any | None = None, verbose: bool = False) -> Callable: ...
